requirements.txt
pyrogram
tgcrypto
requests
pillow
rembg
motor
python-dotenv
